backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
sqlalchemy>=2.0,<3
psycopg[binary]>=3.1,<4
pydantic>=2.7,<3
pydantic-settings>=2.3,<3
alembic>=1.13,<2
playwright>=1.45,<2
celery>=5.3,<6
redis>=5.0,<6
python-multipart>=0.0.9,<1
pdfplumber>=0.11,<1
beautifulsoup4>=4.12,<5
httpx>=0.27,<1
python-jose[cryptography]>=3.3,<4
email-validator>=2.0.0,<3
